SYNAPSE_extracted/SYNAPSE/project/Synapse/synapse/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# ── Django Core ──────────────────────────────────────────────
Django>=4.2,<5.0
# ── WebSockets / Real-time ────────────────────────────────────
channels>=4.1,<5.0
daphne>=4.1,<5.0
channels-redis>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
django-axes>=6.4,<7.0
django-filter>=24.2,<25.0
django-redis>=5.4,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0
django-prometheus>=2.3,<3.0
django-storages[s3]>=1.14,<2.0
django-silk>=5.1,<6.0
whitenoise[brotli]>=6.7,<7.0

# ── Security ─────────────────────────────────────────────────
cryptography>=42.0,<43.0
django-ratelimit>=4.1,<5.0
# Phase 9.1 — MFA
django-otp>=1.4,<2.0
qrcode[pil]>=7.4,<8.0
# Phase 9.1 — Secret scanning (dev/CI)
semgrep>=1.70,<2.0

# ── Database ─────────────────────────────────────────────────
psycopg2-binary>=2.9,<3.0
pgvector>=0.2,<1.0

# ── Cache / Queue ────────────────────────────────────────────
redis>=5.0,<6.0
celery>=5.4,<6.0
flower>=2.0,<3.0

# ── AI / ML ──────────────────────────────────────────────────
langchain==0.3.28
langchain-community==0.3.31
langchain-openai==0.2.14
langchain-google-genai==2.1.12
langchain-core==0.3.83
langgraph>=0.1,<0.3
langsmith>=0.1,<0.3
spacy>=3.7,<4.0
transformers>=4.40,<5.0
accelerate>=0.30,<1.0
sentence-transformers>=3.0,<4.0
torch>=2.3,<3.0
keybert>=0.8,<1.0
rouge-score>=0.1,<1.0
yake>=0.4,<1.0
langdetect>=1.0,<2.0
datasketch>=1.6,<2.0

# ── Web Scraping ─────────────────────────────────────────────
Scrapy>=2.11,<3.0
beautifulsoup4>=4.12,<5.0
playwright>=1.44,<2.0
httpx>=0.27,<1.0
requests>=2.32,<3.0
feedparser>=6.0,<7.0
newspaper3k>=0.2,<1.0
trafilatura>=1.10,<2.0
html2text>=2024.2,<2025.0
fake-useragent>=1.5,<2.0

# ── Document Generation ───────────────────────────────────────
reportlab>=4.2,<5.0
pymupdf>=1.24,<2.0
python-pptx>=0.6,<1.0
python-docx>=1.1,<2.0
fpdf2>=2.7,<3.0
Pillow>=10.3,<11.0
Jinja2>=3.1,<4.0
matplotlib>=3.9,<4.0

# ── Cloud ────────────────────────────────────────────────────
boto3>=1.34,<2.0
google-api-python-client>=2.130,<3.0
google-auth-httplib2>=0.2,<1.0
google-auth-oauthlib>=1.2,<2.0

# ── Email ────────────────────────────────────────────────────
sendgrid>=6.11,<7.0
django-sendgrid-v5>=1.2,<2.0

# ── Monitoring ───────────────────────────────────────────────
sentry-sdk[django]>=2.5,<3.0
prometheus-client>=0.20,<1.0
structlog>=24.2,<25.0

# ── Analytics (Phase 9.2) ────────────────────────────────────
posthog>=3.5,<4.0

# ── Billing (Phase 9.3) ──────────────────────────────────────
stripe>=9.9,<10.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0,<2.0
Markdown>=3.6,<4.0
pydantic>=2.7,<3.0
gunicorn>=22.0,<23.0
isodate>=0.6,<1.0
python-dateutil>=2.9,<3.0
yt-dlp>=2024.1.1

# TASK-504-B1: OpenTelemetry (optional — only loaded if OTEL_ENABLED=true)
# opentelemetry-distro>=0.44b0
# opentelemetry-exporter-otlp-proto-grpc>=1.23.0
# opentelemetry-instrumentation-django>=0.44b0
# opentelemetry-instrumentation-celery>=0.44b0
# opentelemetry-instrumentation-psycopg2>=0.44b0
# opentelemetry-instrumentation-redis>=0.44b0

# TASK-601-B4: PDF export
firebase-admin>=7.0.0
